+++ HITSBlazor/Models/Ideas/Entities/Rating.cs
@@ -8,11 +8,11 @@
         public string ExpertFirstName { get; set; } = string.Empty;
         public string ExpertLastName { get; set; } = string.Empty;
 
-        public int? MarketValue { get; set; }
-        public int? Originality { get; set; }
-        public int? TechnicalRealizability { get; set; }
-        public int? Suitability { get; set; }
-        public int? Budget { get; set; }
+        public byte? MarketValue { get; set; }
+        public byte? Originality { get; set; }
+        public byte? TechnicalRealizability { get; set; }
+        public byte? Suitability { get; set; }
+        public byte? Budget { get; set; }
         public double? RatingValue { get; set; }
         public bool IsConfirmed { get; set; }
