+++ HITSBlazor/Pages/Ideas/IdeasCreate/IdeasCreateModel.cs
@@ -11,13 +11,13 @@ namespace HITSBlazor.Pages.Ideas.IdeasCreate
         public string Result { get; set; } = string.Empty;
         public IdeaStatusType Status { get; set; }
 
-        public int? MaxTeamSize { get; set; }
-        public int? MinTeamSize { get; set; }
+        public byte? MaxTeamSize { get; set; }
+        public byte? MinTeamSize { get; set; }
 
         public string Customer { get; set; } = string.Empty;
         public string ContactPerson { get; set; } = string.Empty;
 
-        public int Suitability { get; set; }
-        public int Budget { get; set; }
+        public byte Suitability { get; set; }
+        public byte Budget { get; set; }
     }
 }
